+++ doubleml/double_ml_pliv.py
@@ -9,7 +9,7 @@ import warnings
 from functools import wraps
 
 from .double_ml import DoubleML
-from ._double_ml_score_mixins import LinearScoreMixin
+from .double_ml_score_mixins import LinearScoreMixin
 from ._utils import _dml_cv_predict, _dml_tune, _check_finite_predictions
